% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates a map plot that shows the location where traffic accidents occurred}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numerical vector corresponding to a particular state in the USA.}

\item{year}{A numerical vector corresponding to a particular year.}
}
\value{
A plot of the map of the state of interest with points on the map
indicating traffics fatalities occurred.
}
\description{
\code{fars_map_state} takes a numerical vector corresponding to the state whose
  data will be plotted and the year when the data occurred. For this function to
  work, data should be in .csv.bz2 format located in the same directory as the
  working directory. In case the data for a particular year is unavailable, the
  function will show an error message for the year that was not available.
}
\details{
If you have a file from FARS that is in .csv format but not in .csv.bz2
  format, you can convert it to .csv.bz2 format by using the write.csv function
  in R. See vignette "Other functions in readFARS".
}
\note{
In case the chosen numerical vector does not correspond to a particular
  state, the function will stop and show an error message. The numbers 3, 7, 14, and any
  number above 56 will result in an error. In case, there are no traffic accidents for that
  particular year or state, a message will appear stating there are no accidents to plot.
}
\examples{
\dontrun{fars_map_state(2, 2013)}
\dontrun{fars_map_state(2000, 2013)}

}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read}}
}
