% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Reading a .csv  or .csv.bz2 file into R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A chararcter string denoting the .csv or .csv.bz2 file to be read into R.}
}
\value{
an R object of class "tbl_df", "tbl", "data.frame".
}
\description{
\code{fars_read} This function is a wrapper fro the read_csv function in
  the package readr. It reads in a single csv file into R for analysis. For
  this to work the file should be a .csv or a .csv.bz2 file located in the same
  directory as the working directory and the package readr and dplyr should be
  installed.
}
\examples{
\dontrun{fars_read("accident_2013.csv")}
\dontrun{fars_read("accident_2014.csv")}
\dontrun{fars_read("accident_2015.csv")}

}
