% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Creates a list of dataframe showing the variables MONTH and year in the data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numerical vector indicating which years to search for in the data.}
}
\value{
A list of data frames for each year that was requested. The data frame
 will have the variables MONTH and year. Each row will contain a single case or
 observation that was recorded for a particular month of that year.
}
\description{
\code{fars_read_years} takes a numerical vector representing which years in the
  data should be included in the data frame output. For this function to work,
  data should be in .csv.bz2 format located in the same directory as the working
  directory. In case the data for a particular year is unavailable, the function
  will show a warning and return null for the year that was not available.
}
\details{
If you have a file from FARS that is in .csv format but not in .csv.bz2
  format, you can convert it to .csv.bz2 format by using the write.csv function.
  in R.
}
\examples{
\dontrun{fars_read_years((c(2013, 2014, 2015)))} ### returns a list of data frames
  ### containing the variables MONTH and year for each year that data is available
\dontrun{fars_read_years(c(2013, 2014, "nineteen ninety two"))} ### returns a list
  ### of data frames containing the variable MONTH and year for the year 2013 and 2014
  ### and an error indicating that "nineteen ninety two" is an invalid request

}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read_years}}
}
