% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a data frame that contains number of records available per month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numerical vector indicating which years to include in the output dataframe.}
}
\value{
A data frame with the variables MONTH and each year where data is
  avaiable, summarizing the number of observations recorded each month per year.
}
\description{
\code{fars_summarize_years} takes a numerical vector representing the years in
  the data  that should be included in the data frame output. For this function
  to work, data should be in .csv.bz2 format located in the same directory as
  the working directory. In case the data for a particular year is unavailable,
  the function will show an error message for the year that was not available.
}
\details{
If you have a file from FARS that is in .csv format but not in .csv.bz2
  format, you can convert it to .csv.bz2 format by using the write.csv function
  in R. See vignette "Other functions in readFARS".
}
\examples{
\dontrun{fars_summarize_years(c(2013, 2014, 2015))} ### returns a data frame
  ### containing the variables MONTH and year that data is available.
\dontrun{fars_summarize_years(c(2013, 2014, "nineteen ninety two"))} ### returns a list
  ### of data frames containing the variable MONTH and year for the year 2013 and 2014
  ### and an error indicating that "nineteen ninety two" is an invalid request.

}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read_years}}
}
