% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARSread.R
\docType{package}
\name{FARSread}
\alias{FARSread}
\alias{FARSread-package}
\title{Facilitate summary and visualization of Fatality Analysis Reporting System (FARS) data}
\description{
FARSread package allows you to create monthly summary tables of accidents per year and
create map plots using geospatial locationing. See the README on
\href{https://github.com/DocOfi/FARSread-Package}{GitHub}
}
\details{
The function you're likely to find most useful from \pkg{FARSread} is
\code{\link{fars_map_state}} to visualize location of accidents per State.
Please refer to the vignettes to see other functions available in the package.
}
