% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name}
\usage{
make_filename(year)
}
\arguments{
\item{filename}{A numerical vector indicating the year to include in the name of 
the file to be created.}
}
\value{
A character vector indicating the name of the file.
}
\description{
\code{make_filename} is a wrapper for the function sprintf that returns a character 
  vector containing a combination of text and the year the data was observed in 
  csv.bz2 format.
}
\examples{
make_filename(1988) ## returns "accident_1988.csv.bz2"
\dontrun{make_filename("Nineteen ninety two")} ## results in error year has to 
  be numeric or an integer
  
}
