% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numerical vector for the year to include in the file name.}
}
\value{
A character vector indicating the name of the file.
}
\description{
\code{make_filename} is a wrapper for the function sprintf that returns a character
  vector containing a combination of text and the year the data was observed in
  csv.bz2 format.
}
